a suggested refactor
