+++ model_run.py
@@ -19,13 +19,12 @@ from time import time
 
 from model_data import (build_source, clean_data, alpha_min)
 from model_factory import polling_model_factory
-from model_solver import solve_model
+from model_solver import solve_model, incorporate_penalties
 from model_results import (
     incorporate_result,
     demographic_domain_summary,
     demographic_summary,
-    write_results,
-    compute_kp_score
+    write_results
 )
 
 CURRENT_DIR = os.path.dirname(os.path.abspath(__file__))
@@ -66,8 +65,16 @@ def run_on_config(config: PollingModelConfig, log: bool=False):
     #incorporate result into main dataframe
     result_df = incorporate_result(dist_df, ea_model)
 
-    #incorporate site penalties as appropriate
-    result_df = incorporate_penalties(dist_df, alpha, run_prefix, result_df, ea_model, config, log)
+    #if this a run with penalized sites, incorporate site penalties 
+    if config.penalized_sites:
+        # if some of the the penalized sites were selected by model 1 (no penalties), run penalty algorithm
+        selected_sites = set(result_df.id_dest)
+        penalized_selections = {x for x in selected_sites if x in config.penalized_sites}
+        if  penalized_selections:
+            print('Penalized sites chosen. Running penalized optimizaion')
+            ea_model_penalized = incorporate_penalties(dist_df, alpha, run_prefix, result_df, ea_model, config, log)
+            #incorporate penalized results
+            result_df = incorporate_result(dist_df, ea_model_penalized)
 
     #calculate the new alpha given this assignment
     alpha_new = alpha_min(result_df)
@@ -94,189 +101,13 @@ def run_on_config(config: PollingModelConfig, log: bool=False):
 
     return result_folder
 
-def incorporate_penalties(dist_df, alpha, run_prefix, result_df, ea_model, config: PollingModelConfig, log: bool=False):
 
 
-    # 0. if there are not penalized sites, we're done
-    if not config.penalized_sites:
-        return result_df
+
     
-    # 1. if none of the penalized sites were selected by model 1 (no penalties), we're done
-    selected_sites = set(result_df.id_dest)
-    penalized_selections = {x for x in selected_sites if x in config.penalized_sites}
-    if not penalized_selections:
-        print('No penalized sites selected')
-        return result_df
-    
-    # otherwise, start a log and continue the algorithm
-    penalty_log = open(get_log_path(config,'penalty'), 'a')
-    penalty_log.write('Model 1: original model with no penalties\n')
-    penalty_log.write(f'Selected {len(penalized_selections)} penalized sites:\n')
-    for s in sorted(penalized_selections):
-        penalty_log.write(f'\t ---> {s}\n')
-
-    # 2. convert objective value to KP score (kp1)
-    obj_value = pyo.value(ea_model.obj)
-    kp1 = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value   
-    penalty_log.write(f'KP Objective = {kp1:.2f}\n')
-
-    # 3. compute optimal solution excluding penalized sites (model 2)
-    ea_model_exclusions = polling_model_factory(dist_df, alpha, config, exclude_penalized_sites=True)
-    if log:
-        print(f'Model 2 (excludes penalized sites) built for {run_prefix}')
-
-    solve_model(ea_model_exclusions, config.time_limit, log=log, log_file_path=get_log_path(config,'model2'))
-    if log:
-        print(f'Model 2 solved for {run_prefix}.')
-
-    # 4. convert objective value to KP score (kp2)
-    obj_value = pyo.value(ea_model_exclusions.obj)
-    kp2 = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value
-    
-    # 5. compute penalty as (kp2-kp1)/len(selected penalized sites in model 1)
-    penalty = (kp2-kp1)/len(penalized_selections)
-    if log:
-        print(f'{kp1 = :.2f}, {kp2 = :.2f}')
-        print(f'computed penalty is {penalty:.2f}')
-    penalty_log.write('\nModel 2: penalized sites excluded\n')
-    penalty_log.write(f'KP Objective = {kp2:.2f}\n')
-     
-    # 6. compute optimal solution including penalized sites applying calculate penalty (model 3)
-    ea_model_penalized =  polling_model_factory(dist_df, alpha, config,
-                                            exclude_penalized_sites=False,
-                                            site_penalty=penalty,
-                                            kp_penalty_parameter=kp1)
-    if log:
-        # final_model = ('penalized', ea_model_penalized)
-        print(f'Model 3 (penalized model) built for {run_prefix}.')
-
-    solve_model(ea_model_penalized, config.time_limit, log=log, log_file_path=get_log_path(config,'model3'))
-    if log:
-        print(f'Model 3 solved for {run_prefix}.')
-
-    # 8. continue to result_df with solution to model 3
-    result_df = incorporate_result(dist_df, ea_model_penalized)
     
 
-    selected_sites = set(result_df.id_dest)
-    penalized_selections = {x for x in selected_sites if x in config.penalized_sites}
-    penalty_log.write('\nModel 3: penalized model\n')
-    penalty_log.write(f'Penalty applied to each site = {penalty:.2f}\n')
-    if penalized_selections:
-        penalty_log.write(f'Selected {len(penalized_selections)} penalized sites:\n')
-        for s in sorted(penalized_selections):
-            penalty_log.write(f'\t ---> {s}\n')
-    else:
-        penalty_log.write('Selected no penalized sites.\n')
-
-    # report some final statistics
-    obj_value = pyo.value(ea_model_penalized.obj)
-    kp_pen = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value
-    kp = compute_kp_score(result_df, config.beta, alpha=alpha)
-    penalty_log.write(f'Penalized KP Optimal = {kp_pen:.2f}\n')
-    penalty_log.write(f'KP Optimal = {kp:.2f}\n')
-    penalty_log.write(f'Penalty = {kp_pen-kp:.2f}\n')
 
 
-    return result_df
 
 
-    # #build model
-    # ea_model = polling_model_factory(dist_df, alpha, config, exclude_penalized_sites=False)
-    # final_model = ('base model', ea_model)
-
-    # line_number = my_print(f'model built for {run_prefix}.', line_number)
-
-    # #solve model
-    # solve_model(ea_model, config.time_limit, log=log, log_file_path=config.log_file_path)
-    # line_number = my_print(f'model solved for {run_prefix}.', line_number)
-
-    # #incorporate result into main dataframe
-    # result_df = incorporate_result(dist_df, ea_model)
-
-    # selected_sites = set(result_df.id_dest)
-    # penalized_selections = {x for x in selected_sites if x in config.penalized_sites}
-
-    # line_number = my_print(f'Unpenalized model selected {len(penalized_selections)} penalized sites:', line_number)
-    # for s in sorted(penalized_selections):
-    #     line_number = my_print(f'\t ---> {s}', line_number)
-
-    # if penalized_selections: # penalized sites were selected, move to step 3
-    #     obj_value = pyo.value(ea_model.obj)
-    #     kp1 = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value
-            
-
-    #     # solve model with penalized sites excluded
-    #     ea_model_exclusions = polling_model_factory(dist_df, alpha, config, exclude_penalized_sites=True)
-    #     line_number = my_print(f'model with exclusions built for {run_prefix}.', line_number)
-
-    #     #solve model
-    #     solve_model(ea_model_exclusions, config.time_limit, log=log, log_file_path=config.log_file_path)
-    #     line_number = my_print(f'model with exclusions solved for {run_prefix}.', line_number)
-        
-    #     obj_value = pyo.value(ea_model_exclusions.obj)
-    #     kp2 = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value
-
-    #     penalty = (kp2-kp1)/len(penalized_selections)
-
-    #     line_number = my_print(f'{kp1 = :.2f}, {kp2 = :.2f}', line_number)
-    #     line_number = my_print(f'computed penalty is {penalty:.2f}', line_number)
-
-    #     ea_model_penalized =  polling_model_factory(dist_df, alpha, config,
-    #                                                 exclude_penalized_sites=False,
-    #                                                 site_penalty=penalty,
-    #                                                 kp_penalty_parameter=kp1)
-    #     final_model = ('penalized', ea_model_penalized)
-    #     line_number = my_print(f'penalized model built for {run_prefix}.', line_number)
-
-    #     #solve model
-    #     solve_model(ea_model_penalized, config.time_limit, log=log, log_file_path=config.log_file_path)
-    #     line_number = my_print(f'penalized model solved for {run_prefix}.', line_number)
-
-    #     #incorporate result into main dataframe
-    #     result_df = incorporate_result(dist_df, ea_model_penalized)
-        
-    #     selected_sites = set(result_df.id_dest)
-    #     penalized_selections = {x for x in selected_sites if x in config.penalized_sites}
-
-    #     if penalized_selections:
-    #         line_number = my_print(f'Penalized model selected {len(penalized_selections)} penalized sites:', line_number)
-    #         for s in sorted(penalized_selections):
-    #             line_number = my_print(f'\t ---> {s}', line_number)
-    #     else:
-    #         line_number = my_print('Penalized model selected no penalized sites.', line_number)
-
-
-    # obj_value = pyo.value(final_model[1].obj)
-    # kp_pen = -1/(config.beta*alpha)*math.log(obj_value) if config.beta else obj_value
-    # kp = compute_kp_score(result_df, config.beta, alpha=alpha)
-    # line_number = my_print(f'Final Model = {final_model[0]}, KP Pen = {kp_pen:.2f}, KP = {kp:.2f}, Penalty={kp_pen-kp:.2f}', line_number)
-
-    # #calculate the new alpha given this assignment
-    # alpha_new = alpha_min(result_df)
-
-    # #calculate the average distances traveled by each demographic to the assigned precinct
-    # demographic_prec = demographic_domain_summary(result_df, 'id_dest')
-
-    # #calculate the average distances traveled by each demographic by residence
-    # demographic_res = demographic_domain_summary(result_df, 'id_orig')
-
-    # #calculate the average distances (and y_ede if beta !=0) traveled by each demographic
-    # demographic_ede = demographic_summary(demographic_res, result_df, config.beta, alpha_new)
-
-    # result_folder = config.result_folder
-
-    # write_results(
-    #     result_folder,
-    #     run_prefix,
-    #     result_df,
-    #     demographic_prec,
-    #     demographic_res,
-    #     demographic_ede,
-    # )
-
-    # return result_folder
-
-def get_log_path(config: PollingModelConfig, specifier: str):
-    base_path = '.'.join(config.log_file_path.split('.')[0:-1])
-    return f'{base_path}.{specifier}.log'
